api/requirements.txt
fastapi[standard]
argon2-cffi
PyJWT
pika
redis